Removed unwanted codes
