+++ Tella/Domain/Entity/UwaziEntity/UwaziValue.swift
@@ -1,0 +1,20 @@
+//
+//  UwaziValue.swift
+//  Tella
+//
+//  Created by Robert Shrestha on 9/13/23.
+//  Copyright © 2023 HORIZONTAL. All rights reserved.
+//
+
+import Foundation
+class UwaziValue {
+    var stringValue: String
+    var selectedValue: [SelectValue]
+    init(stringValue: String, selectedValue: [SelectValue]) {
+        self.stringValue = stringValue
+        self.selectedValue = selectedValue
+    }
+    static func defaultValue() -> UwaziValue {
+        return UwaziValue(stringValue: "", selectedValue: [])
+    }
+}
