+++ Tella/Scenes/Uwazi/Entity/Views/CreateEntityView.swift
@@ -50,7 +50,6 @@ struct CreateEntityView: View {
                 VStack(alignment: .leading) {
                     ForEach(entityViewModel.entryPrompts, id: \.id) { prompt in
                         RenderPropertyComponentView(prompt: prompt)
-                            .environmentObject(entityViewModel)
                     }
                 }.padding(EdgeInsets(top: 16, leading: 16, bottom: 0, trailing: 16))
             }
